titan-poss-poss-web-f78a2c9ef984/titan-poss-poss-web-f78a2c9ef984/libs/poss/shared/view-tcs/data-access-view-tcs/src/lib/+state/view-tcs.selectors.ts
import { createSelector, createFeatureSelector } from '@ngrx/store';
import { selectViewTcsState } from './view-tcs.reducer';

import { ViewTcsState, viewTcsFetureKey } from './view-tcs.state';

export const selectViewTcsaState = createFeatureSelector<ViewTcsState>(
  viewTcsFetureKey
);

export const selectTcsDetails = createSelector(
  selectViewTcsState,
  state => state.tcsDetails
);

export const selectIsLoading = createSelector(
  selectViewTcsState,
  state => state.isLoading
);

export const selectError = createSelector(
  selectViewTcsState,
  state => state.error
);

export const ViewTcsSelectors = {
  selectTcsDetails,
  selectError,
  selectIsLoading
};
